MyInfoBaseV1.0.3.1变更
MyInfoBaseV1.0.3.1变更记录：
1.记住所有数据库保存位置的功能。
2.无信息时的图标类型及显示。
3.程序在IDE中正常运行，但发布后发现程序在节点改名后如果此时迅速添加其他节点会卡死，于是将节点结束编辑后更新数据库的操作放到了独立的线程中。
4.添加双击打开文件的功能。
5.添加在增加文件时，从UI线程跨到呈现线程，显示正在添加的文件功能。
6.添加对称加密与非对称加密解密功能。
7.修复启动应用瞬间还未加载完成树结构时关闭应用导致树结构保存为空的bug。

+++ LabelNode/LabelNodeAccess.cs
@@ -116,7 +116,7 @@ namespace LabelNode
         }
 
 
-        public void DeleteFilesOfInfoNodeDB(String InfoNodeDBPath, List<int> fileIDs)
+        public void DeleteFilesOfInfoNodeDB(String InfoNodeDBPath, List<long> fileIDs)
         {
             repository.DeleteInfoNodesOfLabelNodeDB(InfoNodeDBPath, fileIDs);
         }
